Rename src/tiff.rs to src/geotiff.rs
Reduce confusion when using the `tiff` crate later.
